requirements.txt
Django==5.1.1
psycopg[binary,pool]==3.2.1
djangorestframework==3.15.2
adrf==0.1.7

drf-spectacular==0.27.2
django-environ==0.11.2
Faker==28.4.1

uvicorn==0.30.6
uvicorn-worker==0.2.0
gunicorn==23.0.0
progress==1.6